+++ test/index.test.js
@@ -73,6 +73,248 @@ describe('gridlFactory', () => {
 
     });
 
+    describe('constants', () => {
+
+        describe('adjacences', () => {
+
+            it('should be immutable', () => {
+                expect(() => adjacences.test = 'test').to.throw();
+            });
+
+            describe('ALL', () => {
+
+                it('should return all direct neighbours in the order: left -> right / top -> down', () => {
+                    expect(adjacences.ALL).to.deep.equal([[-1,-1], [0,-1], [1,-1], [-1,0], [1,0], [-1,1], [0,1], [1,1]]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.ALL[0] = 'test').to.throw();
+                });
+
+            });
+
+            describe('ALL_CW', () => {
+
+                it('should return all direct neighbours in clockwise order', () => {
+                    expect(adjacences.ALL_CW).to.deep.equal([
+                        [0,-1],
+                        [1,-1],
+                        [1,0],
+                        [1,1],
+                        [0,1],
+                        [-1,1],
+                        [-1,0],
+                        [-1,-1],
+                    ]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.ALL_CW[0] = 'test').to.throw();
+                });
+
+            });
+
+            describe('ALL_CCW', () => {
+
+                it('should return all direct neighbours in counterclockwise order', () => {
+                    expect(adjacences.ALL_CCW).to.deep.equal([
+                        [0,-1],
+                        [-1,-1],
+                        [-1,0],
+                        [-1,1],
+                        [0,1],
+                        [1,1],
+                        [1,0],
+                        [1,-1],
+                    ]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.ALL_CCW[0] = 'test').to.throw();
+                });
+
+            });
+
+            describe('ORTHOGONAL', () => {
+
+                it('should return all orthogonal direct neighbours in the order: left -> right / top -> bottom', () => {
+                    expect(adjacences.ORTHOGONAL).to.deep.equal([[0,-1], [-1,0], [1,0], [0,1]]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.ORTHOGONAL[0] = 'test').to.throw();
+                });
+
+            });
+
+            describe('ORTHOGONAL_CW', () => {
+
+                it('should return all orthogonal direct neighbours in clockwise order', () => {
+                    expect(adjacences.ORTHOGONAL_CW).to.deep.equal([[0,-1], [1,0], [0,1], [-1,0]]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.ORTHOGONAL_CW[0] = 'test').to.throw();
+                });
+
+            });
+
+            describe('ORTHOGONAL_CCW', () => {
+
+                it('should return all orthogonal direct neighbours in counterclockwise order', () => {
+                    expect(adjacences.ORTHOGONAL_CCW).to.deep.equal([[0,-1], [-1,0], [0,1], [1,0]]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.ORTHOGONAL_CCW[0] = 'test').to.throw();
+                });
+
+            });
+
+            describe('DIAGONAL', () => {
+
+                it('should return all diagonal direct neighbours in the order: left -> right / top -> down', () => {
+                    expect(adjacences.DIAGONAL).to.deep.equal([[-1,-1], [1,-1], [-1,1], [1,1]]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.DIAGONAL[0] = 'test').to.throw();
+                });
+
+            });
+
+            describe('DIAGONAL_CW', () => {
+
+                it('should return all diagonal direct neighbours in clockwise order', () => {
+                    expect(adjacences.DIAGONAL_CW).to.deep.equal([[1,-1], [1,1], [-1,1], [-1,-1]]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.DIAGONAL_CW[0] = 'test').to.throw();
+                });
+
+            });
+
+            describe('DIAGONAL_CCW', () => {
+
+                it('should return all diagonal direct neighbours in counterclockwise order', () => {
+                    expect(adjacences.DIAGONAL_CCW).to.deep.equal([[-1,-1], [-1,1], [1,1], [1,-1]]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => adjacences.DIAGONAL_CCW[0] = 'test').to.throw();
+                });
+
+            });
+
+        });
+
+        describe('directions', () => {
+
+            it('should be immutable', () => {
+                expect(() => directions.test = 'test').to.throw();
+            });
+
+            describe('UP', () => {
+
+                it('should return the expected direction', () => {
+                    expect(directions.UP).to.deep.equal([0,-1]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => directions.UP[0] = 2).to.throw();
+                });
+
+            });
+
+            describe('UP_RIGHT', () => {
+
+                it('should return the expected direction', () => {
+                    expect(directions.UP_RIGHT).to.deep.equal([1,-1]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => directions.UP_RIGHT[0] = 666).to.throw();
+                });
+
+            });
+
+            describe('UP_RIGHT', () => {
+
+                it('should return the expected direction', () => {
+                    expect(directions.RIGHT).to.deep.equal([1,0]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => directions.RIGHT[0] = 666).to.throw();
+                });
+
+            });
+
+            describe('DOWN_RIGHT', () => {
+
+                it('should return the expected direction', () => {
+                    expect(directions.DOWN_RIGHT).to.deep.equal([1,1]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => directions.DOWN_RIGHT[0] = 666).to.throw();
+                });
+
+            });
+
+            describe('DOWN', () => {
+
+                it('should return the expected direction', () => {
+                    expect(directions.DOWN).to.deep.equal([0,1]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => directions.DOWN[0] = 666).to.throw();
+                });
+
+            });
+
+            describe('DOWN_LEFT', () => {
+
+                it('should return the expected direction', () => {
+                    expect(directions.DOWN_LEFT).to.deep.equal([-1,1]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => directions.DOWN_LEFT[0] = 666).to.throw();
+                });
+
+            });
+
+            describe('LEFT', () => {
+
+                it('should return the expected direction', () => {
+                    expect(directions.LEFT).to.deep.equal([-1,0]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => directions.LEFT[0] = 666).to.throw();
+                });
+
+            });
+
+            describe('UP_LEFT', () => {
+
+                it('should return the expected direction', () => {
+                    expect(directions.UP_LEFT).to.deep.equal([-1,-1]);
+                });
+
+                it('should be immutable', () => {
+                    expect(() => directions.UP_LEFT[0] = 666).to.throw();
+                });
+
+            });
+
+        });
+
+    });
+
     describe('validation', () => {
 
         it('should throw an error if data is not an array', () => {