+++ docs/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+**v0.8.6**
+
+* added more predefined adjacence sets
+    * adjacences.ALL_CW
+    * adjacences.ALL_CCW
+    * adjacences.DIAGONAL_CW
+    * adjacences.DIAGONAL_CCW
+    * adjacences.HORIZONTAL_CW
+    * adjacences.HORIZONTAL_CCW
+* made all adjacence and direction constants immutable
+
 **v0.8.5**
 
 * implemented reduce() method